requirements.txt
numpy
pandas
dataframe-image
flask
flask-bootstrap
flask-restful
flask-marshmallow
jinja2
requests
ntplib
pybase64
web3
eth-tester
py-solc==0.4.0
py-solc-x
flask-sqlalchemy
flask-wtf
email_validator
wtforms
flask-bcrypt
Flask-Reuploaded
flask-login
flask-msearch
flask_migrate
flask-uploads
eth-utils
python-dotenv
eth-brownie
pysha3
pycryptodome